project/requirements.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
distlib==0.3.6
Django==4.1.6
django-adminlte-3==0.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
kombu==5.2.4
Pillow==9.5.0
platformdirs==3.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pydantic==1.10.7
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.3
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==2.0.2
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
